+++ src/spond_classes/role.py
@@ -21,4 +21,4 @@ class Role(BaseModel):
 
     def __str__(self) -> str:
         """Return simple human-readable description."""
-        return f"Role(uid='{self.uid}', name='{self.name}')"
+        return f"{self.__class__.__name__}(uid='{self.uid}', name='{self.name}')"