+++ src/spond_classes/subgroup.py
@@ -21,4 +21,4 @@ class Subgroup(BaseModel):
 
     def __str__(self) -> str:
         """Return simple human-readable description."""
-        return f"Subgroup(uid='{self.uid}', name='{self.name}')"
+        return f"{self.__class__.__name__}(uid='{self.uid}', name='{self.name}')"